upgrade constants
