+++ constants/constants.js
@@ -1,0 +1,46 @@
+export const ONBOARDING_SCREENS = [
+  "DesiredResultsScreen",
+  "SelectRoomsScreen",
+  "LivingRoomStatsScreen",
+  "CurrentHomeFeelScreen",
+  "SpaceAffectsMoodScreen",
+  "EighthScreen",
+  "RedesignExperienceScreen",
+  "RedesignChallengesScreen",
+  "EleventhScreen",
+  "FinalRoomsScreen",
+  "ThirteenthScreen",
+  "StyleRecommendationScreen",
+];
+
+export const carouselData = [
+  { id: "1", image: require("./assets/images/Artdeco.png") },
+  { id: "2", image: require("./assets/images/Bauhaus.png") },
+  { id: "3", image: require("./assets/images/Japanese.png") },
+  { id: "4", image: require("./assets/images/Tropical.png") },
+  { id: "5", image: require("./assets/images/Scandinavian.png") },
+];
+
+export const testimonials = [
+  {
+    id: "1",
+    text: "Like Having a Pro at home",
+    maintext:
+      "Recently started remodeling our bedroom and I was stuck for ideas. Tried this app out of curiosity and I'm so glad I did! It's like having a professional help.",
+    name: "Our user",
+  },
+  {
+    id: "2",
+    text: "Like Having a Pro at home",
+    maintext:
+      "Recently started remodeling our bedroom and I was stuck for ideas. Tried this app out of curiosity and I'm so glad I did! It's like having a professional help.",
+    name: "Our user",
+  },
+  {
+    id: "3",
+    text: "Like Having a Pro at home",
+    maintext:
+      "Recently started remodeling our bedroom and I was stuck for ideas. Tried this app out of curiosity and I'm so glad I did! It's like having a professional help.",
+    name: "Our user",
+  },
+];
