+++ constants/modalconstants.js
@@ -1,0 +1,75 @@
+export const room = [
+  { name: "Living room" },
+  { name: "Bedroom" },
+  { name: "Bathroom" },
+  { name: "Kitchen" },
+  { name: "Dining room" },
+  { name: "Attic" },
+  { name: "Study" },
+  { name: "Kids room" },
+  { name: "Formal dining room" },
+  { name: "Home theatre" },
+  { name: "Balcony" },
+  { name: "Workshop" },
+  { name: "Gym" },
+  { name: "Coffee shop" },
+  { name: "Restaurant" },
+];
+export const rooms = [
+  {
+    name: "Soft",
+    description: "Green, sunny, colorful.",
+  },
+  {
+    name: "Balanced",
+    description: "Green, sunny, colorful.",
+  },
+  {
+    name: "Strong",
+    description: "Green, sunny, colorful.",
+  },
+  {
+    name: "Extreme",
+    description: "Green, sunny, colorful.",
+  },
+];
+export const stylesList = [
+  {
+    name: "Artdeco",
+    image: require("./assets/images/Artdeco.png"),
+    description: "Bright, airy, clean",
+  },
+  {
+    name: "Bauhaus",
+    image: require("./assets/images/Bauhaus.png"),
+    description: "Minimal, clear, purposeful",
+  },
+  {
+    name: "Futuristic",
+    image: require("./assets/images/Futuristic.png"),
+    description: "Bright, airy, clean",
+  },
+  {
+    name: "Japanese",
+    image: require("./assets/images/Japanese.png"),
+    description: "Bright, airy, clean",
+  },
+  {
+    name: "Scandinavian",
+    image: require("./assets/images/Scandinavian.png"),
+    description: "Bright, airy, clean",
+  },
+  {
+    name: "Tropical",
+    image: require(".//assets/images/Tropical.png"),
+    description: "Green, sunny, colorful.",
+  },
+];
+export const styleImages = {
+  Artdeco: require("./assets/images/Artdeco.png"),
+  Bauhaus: require("./assets/images/Bauhaus.png"),
+  Futuristic: require("./assets/images/Futuristic.png"),
+  Japanese: require("./assets/images/Japanese.png"),
+  Scandinavian: require("./assets/images/Scandinavian.png"),
+  Tropical: require(".//assets/images/Tropical.png"),
+};
