+++ screeens/Onboarding/ResponseScreen.js
@@ -8,9 +8,9 @@ import {
   TouchableOpacity,
 } from "react-native";
 import FontAwesome from "@expo/vector-icons/FontAwesome";
-import { ONBOARDING_SCREENS } from "../../constants";
+import { ONBOARDING_SCREENS, testimonials } from "../../constants/constants";
 import OnboardingDots from "../../components/Dots";
-import { testimonials } from "../../constants";
+
 import UniversalButton from "../../components/ContinueButtons/UniversalButton";
 const { width } = Dimensions.get("window");
 
@@ -187,7 +187,7 @@ const styles = StyleSheet.create({
     paddingHorizontal: 30,
 
     elevation: 2,
-    height: 205, 
+    height: 205,
   },
 
   reviewText: {
