+++ screeens/Onboarding/SelectStyleScreen.js
@@ -9,8 +9,8 @@ import {
   ImageBackground,
 } from "react-native";
 import FontAwesome from "@expo/vector-icons/FontAwesome";
-import { stylesList } from "../../constants";
-import { ONBOARDING_SCREENS } from "../../constants";
+import { stylesList } from "../../constants/constants";
+import { ONBOARDING_SCREENS } from "../../constants/constants";
 import OnboardingDots from "../../components/Dots";
 const currentIndex = ONBOARDING_SCREENS.indexOf("ThirteenthScreen");
 const ThirteenthScreen = ({ navigation }) => {
